In this updated version, the Scanner class is used to read input from the user via the command line.
The user is prompted to enter the value of x and y, and these values are stored in the int variables x and y, respectively. The sum of x and y is then calculated and stored in the int variable ans. Finally, the result is printed to the console using System.out.println().